check with pkg-config exists before install

+++ build_deps.sh
@@ -73,35 +73,43 @@ else
 fi
 
 #*****************************************************************************install Nlopt
-cd /tmp
-wget http://ab-initio.mit.edu/nlopt/nlopt-2.4.2.tar.gz -O - | tar -zx
-cd nlopt-2.4.2
-./configure --enable-shared --without-guile --prefix=$PREFIX
-make -j || make && make install
-
+if ! pkg-config --exists --atleast-version=2.4 nlopt
+then
+    cd /tmp
+    wget http://ab-initio.mit.edu/nlopt/nlopt-2.4.2.tar.gz -O - | tar -zx
+    cd nlopt-2.4.2
+    ./configure --enable-shared --without-guile --prefix=$PREFIX
+    make -j || make && make install
+fi
 
 #*****************************************************************************install gsl
-cd /tmp
-wget https://ftp.gnu.org/gnu/gsl/gsl-2.4.tar.gz -O - | tar -zx
-cd gsl-2.4
-./configure --prefix=$PREFIX
-make -j || make && make install
-
+if ! pkg-config --exists --atleast-version=2.4 gsl
+then
+    cd /tmp
+    wget https://ftp.gnu.org/gnu/gsl/gsl-2.4.tar.gz -O - | tar -zx
+    cd gsl-2.4
+    ./configure --prefix=$PREFIX
+    make -j || make && make install
+fi
 #*****************************************************************************install hwloc
-cd /tmp
-wget https://www.open-mpi.org/software/hwloc/v1.11/downloads/hwloc-1.11.5.tar.gz -O - | tar -zx
-cd hwloc-1.11.5
-./configure --prefix=$PREFIX
-make -j || make && make install
-
+if ! pkg-config --exists --atleast-version=1.11 hwloc
+then
+    cd /tmp
+    wget https://www.open-mpi.org/software/hwloc/v1.11/downloads/hwloc-1.11.5.tar.gz -O - | tar -zx
+    cd hwloc-1.11.5
+    ./configure --prefix=$PREFIX
+    make -j || make && make install
+fi
 #*****************************************************************************install Starpu
-cd /tmp
-wget http://starpu.gforge.inria.fr/files/starpu-1.2.5/starpu-1.2.5.tar.gz
-tar -zxvf starpu-1.2.5.tar.gz
-cd starpu-1.2.5
-./configure -disable-cuda -disable-mpi --disable-opencl --prefix=$PREFIX
-make -j || make && make install
-
+if ! pkg-config --exists --atleast-version=1.2 starpu
+then
+    cd /tmp
+    wget http://starpu.gforge.inria.fr/files/starpu-1.2.5/starpu-1.2.5.tar.gz
+    tar -zxvf starpu-1.2.5.tar.gz
+    cd starpu-1.2.5
+    ./configure -disable-cuda -disable-mpi --disable-opencl --prefix=$PREFIX
+    make -j || make && make install
+fi
 #************************************************************************ Install Chameleon - Stars-H - HiCMA 
 cd /tmp && rm -rf /tmp/exageostatR
 git clone https://github.com/ecrc/exageostatR.git
@@ -125,13 +133,13 @@ cd hicma
 mkdir -p build && cd build
 cmake .. -DBUILD_SHARED_LIBS=ON
 make -j || make && make install
-cd /tmp
-cd exageostatR && cd src
-mkdir -p build && cd build
-export CPATH=$CPATH:/tmp/exageostatR/src/hicma/chameleon/coreblas/include/coreblas
-cmake ..
-make -j >/dev/null 2>&1 || make VERBOSE=1
-make install
+#cd /tmp
+#cd exageostatR && cd src
+#mkdir -p build && cd build
+#export CPATH=$CPATH:/tmp/exageostatR/src/hicma/chameleon/coreblas/include/coreblas
+#cmake ..
+#make -j >/dev/null 2>&1 || make VERBOSE=1
+#make install
